Move some important bits of description up
The description of how an assertion is constructed seemed to imply that
the object with the fingerprints was what went in a=identity.  You had
to read much further down to learn this.  This just moves some of the
critical text up, duplicating it really, so that this confusion is less
likely.